// src/modules/music/events/event.types.ts
export enum MusicEventsList {
    play_song = 'playing_song',
    add_song_to_queue = 'add_song_to_queue',
    skip_song = 'skip_song',
    pause_song = 'pause_song',
    resume_song = 'resume_song',
    volume_up = 'volume_up',
    volume_down = 'volume_down',
    //
    playing_song = 'playing_song',
    song_added_to_queue = 'song_added_to_queue',
    skipped_song = 'skipped_song',
    paused_song = 'paused_song',
    resumed_song = 'resumed_song',
    volume_gone_up = 'volume_gone_up',
    volume_gone_down = 'volume_gone_down',
    playing_error = 'playing_error'
}